+++ crates/biome_js_analyze/src/lint/suspicious/no_redundant_use_strict.rs
@@ -4,10 +4,10 @@ use biome_analyze::{
 };
 use biome_console::markup;
 use biome_js_syntax::{
-    AnyJsClass, JsDirective, JsDirectiveList, JsFunctionBody, JsModule, JsScript,
+    AnyJsClass, JsDirective, JsDirectiveList, JsFileSource, JsFunctionBody, JsModule, JsScript,
 };
 
-use biome_rowan::{declare_node_union, AstNode, BatchMutationExt};
+use biome_rowan::{declare_node_union, AstNode, AstNodeList, BatchMutationExt};
 
 declare_lint_rule! {
  /// Prevents from having redundant `"use strict"`.
@@ -98,8 +98,12 @@ impl AnyNodeWithDirectives {
             AnyNodeWithDirectives::JsScript(script) => script.directives(),
         }
     }
+
+    const fn is_script(&self) -> bool {
+        matches!(self, AnyNodeWithDirectives::JsScript(_))
+    }
 }
-declare_node_union! { pub AnyJsStrictModeNode = AnyJsClass| JsModule | JsDirective  }
+declare_node_union! { pub AnyJsStrictModeNode = AnyJsClass | JsModule | JsDirective  }
 
 impl Rule for NoRedundantUseStrict {
     type Query = Ast<JsDirective>;
@@ -112,6 +116,7 @@ impl Rule for NoRedundantUseStrict {
         if node.inner_string_text().ok()? != "use strict" {
             return None;
         }
+        let file_source = ctx.source_type::<JsFileSource>();
         let mut outer_most: Option<AnyJsStrictModeNode> = None;
         let root = ctx.root();
         match root {
@@ -120,9 +125,19 @@ impl Rule for NoRedundantUseStrict {
                 for n in node.syntax().ancestors() {
                     match AnyNodeWithDirectives::try_cast(n) {
                         Ok(parent) => {
-                            for directive in parent.directives() {
+                            let directives_len = parent.directives().len();
+                            for (index, directive) in parent.directives().into_iter().enumerate() {
                                 let directive_text = directive.inner_string_text().ok()?;
+
                                 if directive_text == "use strict" {
+                                    // if we are analysing a commonjs file, we ignore the first directive that we have at the top, because it's not redundant
+                                    if index + 1 == directives_len
+                                        && parent.is_script()
+                                        && file_source.is_script()
+                                        && outer_most.is_none()
+                                    {
+                                        break;
+                                    }
                                     outer_most = Some(directive.into());
                                     break; // continue with next parent
                                 }